Make LList and Stack more manipulatable.
A little bit of bad practice being produced here. Individual nodes can
be accessed within the Linked List and Stack... Ideally, we don't want
to allow that kind of access. However, to achieve a linear smallest last
ordering algorithm (in terms of |V| + |E|), then we need a quick lookup
to a given vertices location in the degree list. This means having
access to the nodes. We might be able to wrap them in an iterator
though and still get the same effect... Clean up for the future.